Override kernel methods.
